% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{color.bar}
\alias{color.bar}
\title{Color Bar}
\usage{
color.bar(lut, min, max = -min, nticks = 5, ticks = seq(min, max, len
  = nticks), title = "")
}
\arguments{
\item{lut}{This is a hue range that will be used to display the same color gradient 
used to fill in the nodes in the network display}

\item{min}{This is an integer representing the starting point of the labled color gradient. 
It is also used for scaling purposes when setting the dimentions of the color bar}

\item{max}{This is an integer representing the ending point of the labled color gradient. 
It is also used for scaling purposes when setting the dimentions of the color bar}

\item{nticks}{This is the number of tick marks the user would like to be displayed across the color bar}

\item{ticks}{This is a sequence with a length of 'nticks' which ranges from min to max evenly spaced}

\item{title}{This is the title that is desired to be displayed based on what data parameters were inputted}
}
\value{
A horizontal color bar representative of the inputted data
}
\description{
This function creates a customized color bar optimized to be used in network_display.R
}
