% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_cor}
\alias{compute_cor}
\title{Compute the correlation}
\usage{
compute_cor(data_group_2, data_group_1, type_of_cor)
}
\arguments{
\item{data_group_2}{a n*p matrix.}

\item{data_group_1}{a n*p matrix}

\item{type_of_cor}{if NULL, pearson correlation coefficient will be calculated.
Otherwise, a character string "spearman" to calculate spearman correlation
coefficient.}
}
\value{
A list of correlation matrices for both group 1 and group 2
}
\description{
Compute either pearson or spearman correlation coefficient.
}
