% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_edge_weights}
\alias{compute_edge_weights}
\title{Compute edge weights}
\usage{
compute_edge_weights(pvalue_edge_fdr, binary_link)
}
\arguments{
\item{pvalue_edge_fdr}{This is the p-value for edges possibly after multiple testing correction.}

\item{binary_link}{This is the binary edge connection.}
}
\value{
Edge weights.
}
\description{
This function computes edge weights based on p-value for edges with directions.
}
