% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_pvalue_edge}
\alias{compute_pvalue_edge}
\title{Compute p-value for edges}
\usage{
compute_pvalue_edge(p, diff, diff_p, m)
}
\arguments{
\item{p}{This is the number of biomarker candidates.}

\item{diff}{This is the delta correlation or partial correlation matrix.}

\item{diff_p}{This is the permutation result from either permutation_cor or permutation_pc.}

\item{m}{This is the number of permutations.}
}
\value{
p-value for edges.
}
\description{
This function computes p-value for edges based on permutation result.
}
