% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_pvalue_edge_fdr}
\alias{compute_pvalue_edge_fdr}
\title{Compute fdr p-value for edges}
\usage{
compute_pvalue_edge_fdr(p, pvalue_edge)
}
\arguments{
\item{p}{This is the number of biomarker candidates.}

\item{pvalue_edge}{This is p-value for edges from compute_pvalue_edge.}
}
\value{
Adjusted p-value for edges by fdr.
}
\description{
This function computes fdr p-value for edges to adjust for multiple testing.
}
