% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_pvalue_edge_two_side}
\alias{compute_pvalue_edge_two_side}
\title{Compute two sided p-value for edges}
\usage{
compute_pvalue_edge_two_side(pvalue_edge)
}
\arguments{
\item{pvalue_edge}{This is p-value for edges from compute_pvalue_edge.}
}
\value{
Two sided p-value for edges.
}
\description{
This function computes two sided p-value for edges based on pvalue_edge.
}
