% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indeed.R
\name{job_summary}
\alias{job_summary}
\title{Function to extract job summary for a given url}
\usage{
job_summary(url)
}
\arguments{
\item{url}{the url of the job summary}
}
\value{
the text field of the job summary
}
\description{
Function to extract job summary for a given url
}
\examples{
res <- jobsearch("4892433118027808", q = "data scientist", l = "boston, ma", start = 1)
jobsum <- sapply(res$url, job_summary)
}

