% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indeed.R
\name{jobsearch}
\alias{jobsearch}
\title{Function to perform indeed job search}
\usage{
jobsearch(publisher_id, q, l, start)
}
\arguments{
\item{publisher_id}{the publisher id for the web api}

\item{q}{query string}

\item{l}{location}

\item{start}{page number}
}
\value{
a data frame
}
\description{
This function perform job scraping given
query string, location, and page number
}
\examples{
jobsearch("4892433118027808", q = "data scientist", l = "boston, ma", start = 1)
}

