% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Met_Group_GU}
\alias{Met_Group_GU}
\title{Group label.}
\format{A data frame with 1 row and 120 (subjects) columns.}
\usage{
Met_Group_GU
}
\description{
A dataset containing group information CIRR group: 0; HCC group: 1.
}
\keyword{datasets}
