% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Met_GU}
\alias{Met_GU}
\title{GU cirrhosis (CIRR) and GU hepatocellular carcinoma (HCC) data.}
\format{
A dataframe with 39 variables (rows) and 120 subjects (columns).
}
\usage{
Met_GU
}
\description{
A dataset containing the expression levels of 39 metabolites on a total of 120 subjects (HCC: 60; CIRR: 60).
}
\keyword{datasets}
