% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Met_name_GU}
\alias{Met_name_GU}
\title{KEGG ID}
\format{
A dataframe with 39 KEGG ID as rows and 1 column.
}
\usage{
Met_name_GU
}
\description{
A dataset containing the KEGG ID for each metabolite.
}
\keyword{datasets}
