% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_display.R
\name{network_display}
\alias{network_display}
\title{Interactive Network Visualization}
\usage{
network_display(
  result = NULL,
  nodesize = "Node_Degree",
  nodecolor = "Activity_Score",
  edgewidth = FALSE,
  layout = "nice"
)
}
\arguments{
\item{result}{This is the result from calling either non_partial_corr() or partial_corr().}

\item{nodesize}{This parameter determines what the size of each node will represent. The options 
are 'Node_Degree', 'Activity_Score','P_Value' and 'Z_Score'. The title of the resulting 
network will identify which parameter is selected to represent the node size. The default 
is Node_Degree.}

\item{nodecolor}{This parameter determines what color each node will be based on a yellow to 
blue color gradient. The options are 'Node_Degree', 'Activity_Score', 'P_Value', and '
Z_Score'. A color bar will be created based on which parameter is chosen. The default is
Activity_Score.}

\item{edgewidth}{This is a boolean value to indicate whether the edgewidth should be representative 
of the weight connection (TRUE) or not (FALSE). The default is FALSE.}

\item{layout}{Users can choose from a a handful of network visualization templates including:
'nice', 'sphere', 'grid', 'star', and 'circle'. The default is nice.}
}
\value{
An interactive depiction of the network resulting from INDEED functions 
    non_partial_corr() or patial_corr().
}
\description{
An interactive function to assist in the visualization of the result from INDEED 
    functions non_partial_corr() or patial_corr(). The size and the color of each node can be 
    adjusted by users to represent either the Node_Degree, Activity_Score, Z_Score, or P_Value.
    The color of the edge is based on the binary value of either 1 corresponding to a positive 
    correlation depicted as green or a negative correlation of -1 depicted as red. Users also 
    have the option of having the width of each edge be proportional to its weight value. The 
    layout of the network can also be customized by choosing from the options: 'nice', 'sphere', 
    'grid', 'star', and 'circle'. Nodes can be moved and zoomed in on. Each node and edge will 
    display extra information when clicked on. Secondary interactions will be highlighted as 
    well when a node is clicked on.
}
