% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{permutation_thres}
\alias{permutation_thres}
\title{Calculate the positive and negative threshold based on the permutation result}
\usage{
permutation_thres(thres_left, thres_right, p, diff_p)
}
\arguments{
\item{thres_left}{2.5 percent left tails.}

\item{thres_right}{2.5 percent right tails.}

\item{p}{number of biomarker candidates.}

\item{diff_p}{permutation results.}
}
\value{
A list of positive and negative threshold
}
\description{
Calculate the positive and negative threshold based on the permutation result.
}
