% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pre_partial}
\alias{pre_partial}
\title{Data preprocessing for partial correlaton analysis}
\usage{
pre_partial(data = NULL, class_label = NULL, id = NULL)
}
\arguments{
\item{data}{input matrix of expression from all metabolites from all samples}

\item{class_label}{a binary array with 0: group 1; 1: group 2.}

\item{id}{an array of biomolecule ID to label.}
}
\value{
a list of processed data for next step and rho, error curve for group 1 and 2
}
\description{
A method that integrates differential expression (DE) analysis
  and differential network (DN) analysis to select biomarker candidates for
  survival time prediction. pre_partial is the pre-processing step for INDEED
  partial differential analysis
}
\examples{
pre_partial(data=Met_GU,class_label = Met_Group_GU,id=Met_name_GU)
}
