% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvalue_M_GU}
\alias{pvalue_M_GU}
\title{P-values obtained by differential expression (DE) analysis.}
\format{A data frame with 39 rows and 3 variables:
\describe{
  \item{KEGG.ID}{KEGG ID}
  \item{p.value}{p-value}
}}
\usage{
pvalue_M_GU
}
\description{
A dataset containing the p-value for each metabolite obtained through DE analysis.
}
\keyword{datasets}
