% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{scale_range}
\alias{scale_range}
\title{Scale list of numbers}
\usage{
scale_range(x)
}
\arguments{
\item{x}{This is a list of numbers taken form on the columns outputted from calling 
non_partial_corr or patial_corr functions.}
}
\value{
Scaled version of data that fits between 0 to 1.
}
\description{
This function is used to help spread out data values across 0 to 1. 
    This is so that it will be easier to distinguish values later incorporated into the 
    network_display function.
}
