% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_rho_partial}
\alias{select_rho_partial}
\title{Data preprocessing for partial correlaton analysis}
\usage{
select_rho_partial(data = NULL, class_label = NULL, id = NULL,
  error_curve = "YES")
}
\arguments{
\item{data}{input matrix of expression from all metabolites from all samples}

\item{class_label}{a binary array with 0: group 1; 1: group 2.}

\item{id}{an array of biomolecule ID to label.}

\item{error_curve, }{default "YES", option on whether a error curve plot will be
printed, user can choose "YES" and "NO"}
}
\value{
a list of processed data for next step and rho, error curve for group 1 and 2
}
\description{
A method that integrates differential expression (DE) analysis
  and differential network (DN) analysis to select biomarker candidates for
  survival time prediction. select_rho_partial is the pre-processing step for INDEED
  partial differential analysis
}
\examples{
select_rho_partial(data=Met_GU,class_label = Met_Group_GU,id=Met_name_GU)
}
