% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_sig}
\alias{select_sig}
\title{Select biomarker candidates}
\usage{
select_sig(x = NULL, class_label = NULL, id = NULL, partial = NULL,
  method = NULL, p_val = NULL)
}
\arguments{
\item{x}{a data frame consists of data from group 1 and group 2.}

\item{class_label}{a binary array with 0: group 1; 1: group 2.}

\item{id}{an array of biomolecule ID.}

\item{partial}{logical. If TRUE the sparse differential network will be obtained
by using partial correlation. If FALSE, correlation.}

\item{method}{a character string indicating which correlation coefficient is
to be computed. One of "pearson" (default) or "spearman".}

\item{p_val}{a path to either a data frame or matrix that contains p-values
or NULL.}
}
\value{
A .csv file containing the p-value, node degree, and activity score
   for each biomarker candidate
}
\description{
A method that integrates differential expression (DE) analysis
  and differential network (DN) analysis to select biomarker candidates for
  survival time prediction.
}
\examples{
select_sig(Met_GU, Met_Group_GU, Met_name_GU, partial = TRUE)


}
