% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{choose_rho}
\alias{choose_rho}
\title{Draw error curve}
\usage{
choose_rho(data, n_fold, rho)
}
\arguments{
\item{data}{This is a matrix.}

\item{n_fold}{This parameter specifies the n number in n-fold cross_validation.}

\item{rho}{This is the regularization parameter values to be evalueated in terms their errors.}
}
\value{
A list of errors and their corresponding \eqn{log(rho)}.
}
\description{
This function draws error curve using cross-validation.
}
