% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{choose_rho}
\alias{choose_rho}
\title{Draw error curve}
\usage{
choose_rho(data, n_fold, rho)
}
\arguments{
\item{data}{a matrix.}

\item{n_fold}{specify n to n-fold cross_validation.}

\item{rho}{multiple regularization parameter values to be evalueated in terms of errors.}
}
\value{
a list of errors and their corresponding \eqn{log(rho)}
}
\description{
Draw error curve using cross-validation.
}
