% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_cor}
\alias{compute_cor}
\title{Compute the correlation}
\usage{
compute_cor(data_group_2, data_group_1, type_of_cor)
}
\arguments{
\item{data_group_2}{This is a n*p matrix}

\item{data_group_1}{This is a n*p matrix}

\item{type_of_cor}{If this is NULL, pearson correlation coefficient will be calculated.
Otherwise, a character string "spearman" will calculate the spearman correlation coefficient.}
}
\value{
A list of correlation matrices for both group 1 and group 2
}
\description{
This function computes either the pearson or spearman correlation coefficient. 
     This function is used in non_partial_corr.R
}
