% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_dns}
\alias{compute_dns}
\title{Calculate the differential network score}
\usage{
compute_dns(binary_link, z_score)
}
\arguments{
\item{binary_link}{binary correlation matrix with 1 indicating positive correlation and -1
indicating negative correlation for each biomolecular pair.}

\item{z_score}{converted from p-value.}
}
\value{
An activity score associated with each biomarker candidate
}
\description{
Calculate differential network score.
}
