% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_par}
\alias{compute_par}
\title{Compute the partial correlation}
\usage{
compute_par(pre_inv)
}
\arguments{
\item{pre_inv}{This is an inverse covariance matrix.}
}
\value{
An \eqn{n * n} partial correlation matrix
}
\description{
Compute the partial correlation coefficient.
     This function is used in partial_corr.R
}
