% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{compute_pvalue_edge}
\alias{compute_pvalue_edge}
\title{Compute p-value for edges}
\usage{
compute_pvalue_edge(p, diff, diff_p, m)
}
\arguments{
\item{p}{This is the number of biomarker candidates.}

\item{diff}{This is the delta correlation or partial correlation matrix.}

\item{diff_p}{This is the permutation result from either permutation_cor or permutation_pc.}

\item{m}{This is the number of permutations.}
}
\value{
p-value for edges.
}
\description{
This function computes p-value for edges based on permutation result.
}
\details{
#' @title Scale list of numbers
#' @description This function is used to help spread out data values across 0 to 1. 
#'     This is so that it will be easier to distinguish values later incorporated into the 
#'     network_display function. 
#' @param x This is a list of numbers taken form on the columns outputted from calling 
#'     non_partial_corr or patial_corr functions.
#' @return Scaled version of data that fits between 0 to 1.

scale_range <- function(x){x/max(x)}
}
