% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDEED.R
\docType{package}
\name{INDEED}
\alias{INDEED}
\alias{INDEED-package}
\title{INDEED: A package for biomarker candidate prioritization.}
\description{
The INDEED package provides an important functions below:
}
\section{select_rho_partial function}{

select_rho_partial function preprocess data for partical correlation analysis,
the result contains list of preprocessed data and rho values and
error plot for user to choose desired rho value
}

\section{partial_cor function}{

partial_cor function performs partical correlation analysis
based on user input preprocessed list from select_rho_partial step
and the rho choosing method or rho of their choice
and number of permutations (default 1000), p-value is optional,
the result of score table and differential network will be returned
}

\section{non_partial_cor function}{

non_partial_cor function performs correlation analysis based on
user input data, class label, p-value, sample id,
number of permutations, and method (default pearson),
p value is optional,
the result of score table and differential network will be returned
}

