% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_sample_data}
\alias{load_sample_data}
\title{Load the sample data}
\usage{
load_sample_data()
}
\value{
A list of sample data sets
}
\description{
Load the sample data sets from GU cohorts.
}
\examples{
load_sample_data()

}
