% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{loglik_ave}
\alias{loglik_ave}
\title{Create log likelihood error}
\usage{
loglik_ave(data, theta)
}
\arguments{
\item{data}{This is a matrix.}

\item{theta}{This is a precision matrix.}
}
\value{
log likelihood error
}
\description{
This function calculates the log likelihood error.
}
