% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{loglik_ave}
\alias{loglik_ave}
\title{Create log likelihood error function}
\usage{
loglik_ave(data, theta)
}
\arguments{
\item{data}{a matrix or data.frame.}

\item{theta}{a precision matrix.}
}
\value{
log likelihood error function
}
\description{
Calculate log likelihood error function.
}
