% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Met_GU}
\alias{Met_GU}
\title{GU cirrhosis (CIRR) and GU Hepatocellular carcinoma (HCC) combined.}
\format{A data frame with 39 variables as rows and 120 subjects as columns.}
\usage{
Met_GU
}
\description{
A dataset containing the expression levels of 39 metabolites for 120 subjects (CIRR: 60; HCC: 60).
}
\keyword{datasets}
