% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_display.R
\name{network_display}
\alias{network_display}
\title{Network Visualization for INDEED Partial and Non-Partial Correlation}
\usage{
network_display(results = NULL, nodesize = "Node_Degree",
  nodecolor = "Activity_Score", edgewidth = "NO", layout = "nice")
}
\arguments{
\item{results}{This is the result from the calling either partial_cor() or non_partial_cor().}

\item{nodesize}{This parameter determines what the size of each node will represent. The options
are 'Node_Degree', 'Activity_Score', 'Z_Score', and 'P_Value'. The title of the resulting network will
identify which parameter was selected to represent the node size. The default is Node_Degree.}

\item{nodecolor}{This parameter determines what color each node will be, based on a yellow or red color gradient.
The options are 'Node_Degree', 'Activity_Score', 'Z_Score', and 'P_Value'. A color bar will be created
based on which parameter is chosen.}

\item{edgewidth}{This is a 'YES' or 'NO' option as to if the edgewidth should be representative of the weight value
corresponding to the correlation between two nodes.}

\item{layout}{User can choose from a a handful of network visualization templates
including:'nice', 'sphere', 'grid', 'star', and 'circle'.}
}
\value{
A visual dipiction of the network resulting from INDEED functions partial_cor() or non_partial_cor()
}
\description{
A function to assist in the visualization of the results from INDEED functions patial_cor()
and non_partial_cor(). The size and the color of each node can be adjusted by users to represent either
the Node_Degree, Activity_Score, Z_Score, or P_Value. The color of the edge is based on the binary value of
either 1 corresonding to a positive correlation change from group 1 to group 2 dipicted as green or a negative 
correlation change dipicted as red. The
user also has the option of having the width of each edge be proportional to its weight value. The layout of the network
can also be customized by choosing from the options: 'nice', 'sphere', 'grid', 'star', and 'circle'.
}
\examples{
result = non_partial_cor(data = Met_GU, class_label = Met_Group_GU, id = Met_name_GU,
                         method = "pearson", permutation_thres = 0.05, permutation = 1000)
network_display(results = result, layout = 'nice', nodesize = 'Node_Degree',
                nodecolor = 'Activity_Score', edgewidth = 'NO')

}
