% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{permutation_pc}
\alias{permutation_pc}
\title{Permutations to build differential network using partial correlation}
\usage{
permutation_pc(m, p, n_group_1, n_group_2, data_group_1, data_group_2,
  rho_group_1_opt, rho_group_2_opt)
}
\arguments{
\item{m}{number of permutations.}

\item{p}{number of biomarker candidates.}

\item{n_group_1}{number of subjects in group 1.}

\item{n_group_2}{number of subjects in group 2.}

\item{data_group_1}{a \eqn{n*p} matrix or data.frame containing group 1 data.}

\item{data_group_2}{a \eqn{n*p} matrix of data.frame containing group 2 data.}

\item{rho_group_1_opt}{optimal tuning parameter to sparse the differential network for group 1}

\item{rho_group_2_opt}{optimal tuning parameter to sparse the differential network for group 2}
}
\value{
A multi-dimensional matrix that contains the permutation results
}
\description{
A permutation test that randomly permutes the sample labels in distinct
    biological groups for each biomolecule. The difference in paired partial correlation
    is considered significant if it falls into the 2.5% tails on either end of the empirical
    distribution curve.
}
