% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{permutation_thres}
\alias{permutation_thres}
\title{Calculate the positive and negative thresholds based on the permutation result}
\usage{
permutation_thres(thres_left, thres_right, p, diff_p)
}
\arguments{
\item{thres_left}{This is the threshold representing 2.5 percent of the left tail of the 
empirical distributuion curve.}

\item{thres_right}{This is the threshold representing 2.5 percent of the right tail of the 
empirical distributuion curve.}

\item{p}{This is the number of biomarker candidates present.}

\item{diff_p}{This is the permutation result from either permutation_cor or permutation_pc.}
}
\value{
A list of positive and negative thresholds.
}
\description{
This function calculates the positive and negative thresholds based on the 
    permutation result.
}
