% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{pvalue_logit}
\alias{pvalue_logit}
\title{Obtain p-values using logistic regression}
\usage{
pvalue_logit(x, class_label, Met_name)
}
\arguments{
\item{x}{This is a data frame consists of data from group 1 and group 2.}

\item{class_label}{This is a binary array indicating 0 for group 1 and 1 for group 2.}

\item{Met_name}{This is an array of IDs.}
}
\value{
p-values
}
\description{
This function calculates p-values using logistic regression in cases that p-values 
    are not provided.
}
