% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rho_partial.R
\name{select_rho_partial}
\alias{select_rho_partial}
\title{Data preprocessing for partial correlation analysis}
\usage{
select_rho_partial(
  data = NULL,
  class_label = NULL,
  id = NULL,
  error_curve = TRUE
)
}
\arguments{
\item{data}{This is a p*n dataframe that contains the expression levels for all biomolecules and samples.}

\item{class_label}{This is a 1*n dataframe that contains the class label with 0 for group 1 and 1 for group 2.}

\item{id}{This is a p*1 dataframe that contains the ID for each biomolecule.}

\item{error_curve}{This is a boolean value indicating whether to plot the error curve (TRUE) or not (FALSE). 
The default is TRUE.}
}
\value{
A list of processed data for the next step, and an error curve to select optimal rho value
    for graphical lasso.
}
\description{
A method that integrates differential expression (DE) analysis
    and differential network (DN) analysis to select biomarker candidates for
    cancer studies. select_rho_partial is the pre-processing step for INDEED
    partial differential analysis.
}
\examples{
select_rho_partial(data = Met_GU, class_label = Met_Group_GU, id = Met_name_GU, 
    error_curve = TRUE)
}
